requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-multipart==0.0.20
certifi==2025.8.3
anthropic==0.54.0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
bcrypt==4.3.0
alembic==1.16.5
cryptography==45.0.6
psycopg2-binary==2.9.10
twilio==9.6.5
